+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   * Add percentiles() stream function
   * Add has_attribute() stream function
   * Make prn() return the same event
+  * Add null() stream function
 
 0.0.6 2014-08-05
 ================